requirements.txt
torch
torchvision
tensorflow
pytest
pillow
numpy
